change data type for nft

+++ art_gallery/src/main/java/art_gallery/Artwork.java
@@ -5,9 +5,9 @@ public class Artwork {
     private String title;
     private Artist artist;
     private double price;
-    private double nft;
+    private int nft;
 
-    public Artwork(String title, Artist artist, double price, double nft){
+    public Artwork(String title, Artist artist, double price, int nft){
         this.title = title;
         this.artist = artist;
         this.price = price;
